djangoback/requirements.txt
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.4.2
Django==2.2.10
django-cors-headers==3.5.0
django-ratelimit==2.0.0
djangorestframework==3.9.2
gunicorn==19.7.1
idna==2.7
python-decouple==3.1
pytz==2020.4
requests==2.20.0
requests-cache==0.5.2
sqlparse==0.4.1
urllib3==1.24.3
